[refactor] #5 UserRepository 패키지 변경
